requirements.txt
hjson==3.1.0
moviepy==1.0.3
numpy==1.26.4
Pillow==10.3.0
pydub==0.25.1
scipy==1.13.1
